Multi-phasor plot and UI improvements (#3)
* Added a multiphasor plotter

This is a very rough implementation of the multiphasor on one plot.

* Multiphasor plot updated with color

Added a color in scatter plot instead of colormaps and plotted the average as a cross on the scatter plot. Things to do: Clean up the code for multiphasor plot.

* Updated the Multiphasor view

Updated on the color_or_colormap variable for more cleaner implementation

* Updated the tiff writing command

Imasave to imwrite

* Update on the multiphasor plot

Added lifetime values on the plot

* Updated style of multi-phasor selector window

* Cleaned up multi-phasor related code in data window class

* Refactored multi-phasor data windows

* Moved the multi phasor button to the side bar

* Implemented save button for graph

---------

Co-authored-by: Srijan Chakraborty <srijanc.1997@gmail.com>